chore: Some OS specific network stuff moved to compat.h
